+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Dates are in `YYYY-MM-DD` format and versions are in [semantic versioning](http:
 ### Fixed
 
 - Save characters even without name and description to prevent data loss.
+- Typos in documentation and tutorial.
 
 ### Maintenance