+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 We'll note all notable changes in this file, including bug fixes, enhancements, and all closed issues.
 Dates are in `YYYY-MM-DD` format and versions are in [semantic versioning](http://semver.org/) format.
 
+## 0.6.22
+
+### Fixed
+
+- Improved tooltip on concept links.
+
 ## 0.6.21
 
 ### Added